requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
billiard==3.5.0.3
boto3==1.5.1
celery==4.1.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
coreschema==0.0.4
cryptography==2.1.4
culqipy==0.2.6
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.6
django-cors-headers==2.1.0
django-filter==1.0.2
django-filters==0.2.1
django-redis==4.8.0
django-taggit==0.22.1
djangorestframework==3.6.2
drfdocs==0.0.11
Fabric==1.14.0
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
kombu==4.1.0
Markdown==2.6.10
MarkupSafe==1.0
oauthlib==2.0.6
olefile==0.44
paramiko==2.4.0
Pillow==4.1.0
psycopg2==2.7.3.2
pyasn1==0.4.2
pycparser==2.18
pycurl==7.43.0
PyJWT==1.5.3
PyNaCl==1.2.1
python-dateutil==2.6.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.3
raven==6.0.0
requests==2.18.4
requests-oauthlib==0.8.0
social-auth-app-django==2.0.0
social-auth-core==1.5.0
uritemplate==3.0.0
urllib3==1.22
uWSGI==2.0.15
vine==1.1.4
